refactor(web): useChannelStatus hook + 抽 WeixinChannelPanel + 删 ChannelsPage
三处样板压成 1 个 hook:
- 新 hooks/useChannelStatus.ts — 2s 轮询 + busy/error/runAction 三件套,
  替代 FeishuChannelPanel 和 WeixinChannelPanel 各自的 useEffect+useState
  status/busy/error 样板(各 ~50 行)
- FeishuChannelPanel 改用 useChannelStatus(() => api.getFeishuStatus()),
  appId/appSecret/defaultAgentId 状态保留在组件内,7 步 wizard JSX 不动
- WeixinChannelPanel 从 ChannelsPanel.tsx 内联函数(55-247 行)抽出,独立
  成 components/channels/WeixinChannelPanel.tsx,与 FeishuChannelPanel 对称

ChannelsPanel.tsx 从 248 行降到 ~120 行,只剩 CHANNELS 数组 +
PlannedChannelPanel + 顶层 ChannelsPanel 容器。

删除 components/channels/ChannelsPage.tsx — 死代码,grep 全仓只在
自身和 CLAUDE.md 出现,无 import。

CSS class 全保留(channels-shell/channels-list/channels-card/
channels-feishu-wizard 等),E2E 选择器零影响。
